Split Task Manager files for Swift budget
