+++ .github/swift-file-length-budget.tsv
@@ -1,14 +1,14 @@
 # cmux-owned Swift file length budget.
 # Format: max_lines<TAB>relative path
 # Reduce counts as files shrink. CI fails if tracked files exceed this budget.
-18136	CLI/cmux.swift
-16470	Sources/TerminalController.swift
+18537	CLI/cmux.swift
+17008	Sources/TerminalController.swift
 15801	Sources/ContentView.swift
-13765	Sources/AppDelegate.swift
+13772	Sources/AppDelegate.swift
 13685	Sources/Workspace.swift
 13265	Sources/GhosttyTerminalView.swift
 10603	Sources/Panels/BrowserPanel.swift
-8457	Sources/cmuxApp.swift
+8463	Sources/cmuxApp.swift
 7495	Sources/TabManager.swift
 6811	Sources/Panels/BrowserPanelView.swift
 5139	cmuxTests/AppDelegateShortcutRoutingTests.swift
@@ -59,7 +59,7 @@
 729	Sources/MainWindowFocusController.swift
 719	cmuxUITests/BrowserOmnibarSuggestionsUITests.swift
 714	Sources/AppleScriptSupport.swift
-708	Sources/App/MenuBarExtraController.swift
+720	Sources/App/MenuBarExtraController.swift
 683	Sources/SocketControlSettings.swift
 664	Sources/App/ShortcutRoutingSupport.swift
 643	Sources/Panels/BrowserPopupWindowController.swift
