+++ Sources/TaskManagerWindowController.swift
@@ -298,398 +298,3 @@ private struct CmuxTaskManagerRowView: View {
         .opacity(row.isDimmed ? 0.68 : 1)
     }
 }
-
-private struct CmuxTaskManagerSnapshot {
-    static let empty = CmuxTaskManagerSnapshot(
-        rows: [],
-        total: .zero,
-        sampledAt: nil
-    )
-
-    let rows: [CmuxTaskManagerRow]
-    let total: CmuxTaskManagerResources
-    let sampledAt: Date?
-
-    var updatedText: String {
-        guard let sampledAt else {
-            return String(localized: "taskManager.updated.never", defaultValue: "Never")
-        }
-        return CmuxTaskManagerFormat.time(sampledAt)
-    }
-
-    init(rows: [CmuxTaskManagerRow], total: CmuxTaskManagerResources, sampledAt: Date?) {
-        self.rows = rows
-        self.total = total
-        self.sampledAt = sampledAt
-    }
-
-    init(payload: [String: Any]) {
-        let sample = payload["sample"] as? [String: Any] ?? [:]
-        self.sampledAt = CmuxTaskManagerFormat.iso8601Date(sample["sampled_at"] as? String)
-        self.total = CmuxTaskManagerResources(payload["totals"] as? [String: Any] ?? [:])
-
-        var rows: [CmuxTaskManagerRow] = []
-        let windows = payload["windows"] as? [[String: Any]] ?? []
-        for window in windows {
-            Self.appendWindow(window, to: &rows)
-        }
-        self.rows = rows
-    }
-
-    private static func appendWindow(_ window: [String: Any], to rows: inout [CmuxTaskManagerRow]) {
-        let handle = displayHandle(window)
-        var detailParts: [String] = []
-        if bool(window["key"]) {
-            detailParts.append(String(localized: "taskManager.row.keyWindow", defaultValue: "Key window"))
-        }
-        if bool(window["visible"]) == false {
-            detailParts.append(String(localized: "taskManager.row.hidden", defaultValue: "Hidden"))
-        }
-        rows.append(row(
-            window,
-            kind: .window,
-            level: 0,
-            title: String(localized: "taskManager.row.window", defaultValue: "Window \(handle)"),
-            detail: detailParts.joined(separator: " / ")
-        ))
-
-        let workspaces = window["workspaces"] as? [[String: Any]] ?? []
-        for workspace in workspaces {
-            appendWorkspace(workspace, to: &rows)
-        }
-    }
-
-    private static func appendWorkspace(_ workspace: [String: Any], to rows: inout [CmuxTaskManagerRow]) {
-        let title = nonEmptyString(workspace["title"]) ?? displayHandle(workspace)
-        var detailParts: [String] = []
-        if bool(workspace["selected"]) {
-            detailParts.append(String(localized: "taskManager.row.selected", defaultValue: "Selected"))
-        }
-        if bool(workspace["pinned"]) {
-            detailParts.append(String(localized: "taskManager.row.pinned", defaultValue: "Pinned"))
-        }
-        rows.append(row(
-            workspace,
-            kind: .workspace,
-            level: 1,
-            title: title,
-            detail: detailParts.joined(separator: " / ")
-        ))
-
-        let tags = workspace["tags"] as? [[String: Any]] ?? []
-        for tag in tags {
-            appendTag(tag, to: &rows)
-        }
-
-        let panes = workspace["panes"] as? [[String: Any]] ?? []
-        for pane in panes {
-            appendPane(pane, to: &rows)
-        }
-    }
-
-    private static func appendTag(_ tag: [String: Any], to rows: inout [CmuxTaskManagerRow]) {
-        let key = nonEmptyString(tag["key"]) ?? String(localized: "taskManager.row.unknownTag", defaultValue: "Unknown tag")
-        let value = nonEmptyString(tag["value"])
-        let title = value.map { "\(key): \($0)" } ?? key
-        let detail = int(tag["pid"]).map {
-            String(localized: "taskManager.row.pid", defaultValue: "PID \($0)")
-        } ?? ""
-        rows.append(row(tag, kind: .tag, level: 2, title: title, detail: detail, isDimmed: bool(tag["visible"]) == false))
-
-        let processes = tag["processes"] as? [[String: Any]] ?? []
-        for process in processes {
-            appendProcess(process, level: 3, to: &rows)
-        }
-    }
-
-    private static func appendPane(_ pane: [String: Any], to rows: inout [CmuxTaskManagerRow]) {
-        let handle = displayHandle(pane)
-        rows.append(row(
-            pane,
-            kind: .pane,
-            level: 2,
-            title: String(localized: "taskManager.row.pane", defaultValue: "Pane \(handle)"),
-            detail: bool(pane["focused"]) ? String(localized: "taskManager.row.focused", defaultValue: "Focused") : ""
-        ))
-
-        let surfaces = pane["surfaces"] as? [[String: Any]] ?? []
-        for surface in surfaces {
-            appendSurface(surface, to: &rows)
-        }
-    }
-
-    private static func appendSurface(_ surface: [String: Any], to rows: inout [CmuxTaskManagerRow]) {
-        let type = (nonEmptyString(surface["type"]) ?? "unknown").lowercased()
-        let title = nonEmptyString(surface["title"]) ?? displayHandle(surface)
-        var detailParts = [surfaceTypeLabel(type)]
-        if bool(surface["selected"]) {
-            detailParts.append(String(localized: "taskManager.row.selected", defaultValue: "Selected"))
-        }
-        if let tty = nonEmptyString(surface["tty"]) {
-            detailParts.append(tty)
-        }
-        if let url = nonEmptyString(surface["url"]) {
-            detailParts.append(url)
-        }
-        rows.append(row(
-            surface,
-            kind: type == "browser" ? .browserSurface : .terminalSurface,
-            level: 3,
-            title: title,
-            detail: detailParts.joined(separator: " / ")
-        ))
-
-        let webviews = surface["webviews"] as? [[String: Any]] ?? []
-        if webviews.isEmpty {
-            let processes = surface["processes"] as? [[String: Any]] ?? []
-            for process in processes {
-                appendProcess(process, level: 4, to: &rows)
-            }
-        } else {
-            for webview in webviews {
-                appendWebView(webview, to: &rows)
-            }
-        }
-    }
-
-    private static func appendWebView(_ webview: [String: Any], to rows: inout [CmuxTaskManagerRow]) {
-        let title = nonEmptyString(webview["title"])
-            ?? String(localized: "taskManager.row.webview", defaultValue: "WebView")
-        var detailParts: [String] = []
-        if let pid = int(webview["pid"]) {
-            detailParts.append(String(localized: "taskManager.row.pid", defaultValue: "PID \(pid)"))
-        }
-        if let sharedCount = int(webview["shared_process_count"]), sharedCount > 1 {
-            detailParts.append(String(localized: "taskManager.row.sharedProcess", defaultValue: "Shared x\(sharedCount)"))
-        }
-        if let url = nonEmptyString(webview["url"]) {
-            detailParts.append(url)
-        }
-        rows.append(row(webview, kind: .webview, level: 4, title: title, detail: detailParts.joined(separator: " / ")))
-
-        let processes = webview["processes"] as? [[String: Any]] ?? []
-        for process in processes {
-            appendProcess(process, level: 5, to: &rows)
-        }
-    }
-
-    private static func appendProcess(_ process: [String: Any], level: Int, to rows: inout [CmuxTaskManagerRow]) {
-        let pid = int(process["pid"])
-        let title = nonEmptyString(process["name"])
-            ?? pid.map { String(localized: "taskManager.row.processWithPID", defaultValue: "Process \($0)") }
-            ?? String(localized: "taskManager.row.process", defaultValue: "Process")
-        let detail = pid.map {
-            String(localized: "taskManager.row.pid", defaultValue: "PID \($0)")
-        } ?? ""
-        rows.append(row(process, kind: .process, level: level, title: title, detail: detail))
-
-        let children = process["children"] as? [[String: Any]] ?? []
-        for child in children {
-            appendProcess(child, level: level + 1, to: &rows)
-        }
-    }
-
-    private static func row(
-        _ payload: [String: Any],
-        kind: CmuxTaskManagerRow.Kind,
-        level: Int,
-        title: String,
-        detail: String,
-        isDimmed: Bool = false
-    ) -> CmuxTaskManagerRow {
-        CmuxTaskManagerRow(
-            id: rowID(payload, kind: kind),
-            kind: kind,
-            level: level,
-            title: title,
-            detail: detail,
-            resources: CmuxTaskManagerResources(payload["resources"] as? [String: Any] ?? [:]),
-            isDimmed: isDimmed
-        )
-    }
-
-    private static func rowID(_ payload: [String: Any], kind: CmuxTaskManagerRow.Kind) -> String {
-        if let id = nonEmptyString(payload["id"]) {
-            return "\(kind.rawValue):\(id)"
-        }
-        if let pid = int(payload["pid"]) {
-            return "\(kind.rawValue):pid:\(pid)"
-        }
-        if let ref = nonEmptyString(payload["ref"]) {
-            return "\(kind.rawValue):\(ref)"
-        }
-        return "\(kind.rawValue):\(UUID().uuidString)"
-    }
-
-    private static func displayHandle(_ payload: [String: Any]) -> String {
-        nonEmptyString(payload["ref"]) ?? nonEmptyString(payload["id"]) ?? "?"
-    }
-
-    private static func surfaceTypeLabel(_ type: String) -> String {
-        switch type {
-        case "browser":
-            return String(localized: "taskManager.row.surfaceType.browser", defaultValue: "Browser")
-        case "terminal":
-            return String(localized: "taskManager.row.surfaceType.terminal", defaultValue: "Terminal")
-        case "unknown", "":
-            return String(localized: "taskManager.row.surfaceType.unknown", defaultValue: "Unknown")
-        default:
-            return type
-        }
-    }
-
-    private static func nonEmptyString(_ raw: Any?) -> String? {
-        guard let value = raw as? String else { return nil }
-        let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
-        return trimmed.isEmpty ? nil : trimmed
-    }
-
-    private static func bool(_ raw: Any?) -> Bool {
-        if let value = raw as? Bool { return value }
-        if let value = raw as? NSNumber { return value.boolValue }
-        return false
-    }
-
-    private static func int(_ raw: Any?) -> Int? {
-        if let value = raw as? Int { return value }
-        if let value = raw as? NSNumber { return value.intValue }
-        if let value = raw as? String {
-            return Int(value.trimmingCharacters(in: .whitespacesAndNewlines))
-        }
-        return nil
-    }
-}
-
-private struct CmuxTaskManagerRow: Identifiable {
-    enum Kind: String {
-        case window
-        case workspace
-        case tag
-        case pane
-        case terminalSurface
-        case browserSurface
-        case webview
-        case process
-
-        var systemImage: String {
-            switch self {
-            case .window: return "macwindow"
-            case .workspace: return "rectangle.stack"
-            case .tag: return "tag"
-            case .pane: return "square.split.2x1"
-            case .terminalSurface: return "terminal"
-            case .browserSurface: return "globe"
-            case .webview: return "network"
-            case .process: return "gearshape"
-            }
-        }
-
-        var tint: Color {
-            switch self {
-            case .window: return .secondary
-            case .workspace: return .accentColor
-            case .tag: return .orange
-            case .pane: return .secondary
-            case .terminalSurface: return .green
-            case .browserSurface: return .blue
-            case .webview: return .purple
-            case .process: return .secondary
-            }
-        }
-    }
-
-    let id: String
-    let kind: Kind
-    let level: Int
-    let title: String
-    let detail: String
-    let resources: CmuxTaskManagerResources
-    let isDimmed: Bool
-}
-
-private struct CmuxTaskManagerResources {
-    static let zero = CmuxTaskManagerResources(cpuPercent: 0, residentBytes: 0, processCount: 0)
-
-    let cpuPercent: Double
-    let residentBytes: Int64
-    let processCount: Int
-
-    init(cpuPercent: Double, residentBytes: Int64, processCount: Int) {
-        self.cpuPercent = cpuPercent
-        self.residentBytes = residentBytes
-        self.processCount = processCount
-    }
-
-    init(_ payload: [String: Any]) {
-        self.cpuPercent = Self.double(payload["cpu_percent"])
-        self.residentBytes = Self.int64(payload["resident_bytes"])
-        self.processCount = Self.int(payload["process_count"]) ?? 0
-    }
-
-    private static func double(_ raw: Any?) -> Double {
-        if let value = raw as? Double { return value }
-        if let value = raw as? NSNumber { return value.doubleValue }
-        if let value = raw as? String,
-           let parsed = Double(value.trimmingCharacters(in: .whitespacesAndNewlines)) {
-            return parsed
-        }
-        return 0
-    }
-
-    private static func int64(_ raw: Any?) -> Int64 {
-        if let value = raw as? Int64 { return value }
-        if let value = raw as? Int { return Int64(value) }
-        if let value = raw as? NSNumber { return value.int64Value }
-        if let value = raw as? String,
-           let parsed = Int64(value.trimmingCharacters(in: .whitespacesAndNewlines)) {
-            return parsed
-        }
-        return 0
-    }
-
-    private static func int(_ raw: Any?) -> Int? {
-        if let value = raw as? Int { return value }
-        if let value = raw as? NSNumber { return value.intValue }
-        if let value = raw as? String {
-            return Int(value.trimmingCharacters(in: .whitespacesAndNewlines))
-        }
-        return nil
-    }
-}
-
-private enum CmuxTaskManagerFormat {
-    private static let isoFormatter = ISO8601DateFormatter()
-    private static let timeFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.timeStyle = .medium
-        formatter.dateStyle = .none
-        return formatter
-    }()
-
-    static func cpu(_ value: Double) -> String {
-        String(format: "%.1f%%", max(0, value))
-    }
-
-    static func bytes(_ bytes: Int64) -> String {
-        let units = ["B", "KB", "MB", "GB", "TB"]
-        var value = Double(max(0, bytes))
-        var unitIndex = 0
-        while value >= 1024, unitIndex < units.count - 1 {
-            value /= 1024
-            unitIndex += 1
-        }
-        if unitIndex == 0 {
-            return "\(Int(value)) \(units[unitIndex])"
-        }
-        return String(format: "%.1f %@", value, units[unitIndex])
-    }
-
-    static func iso8601Date(_ raw: String?) -> Date? {
-        guard let raw else { return nil }
-        return isoFormatter.date(from: raw)
-    }
-
-    static func time(_ date: Date) -> String {
-        timeFormatter.string(from: date)
-    }
-}
